# public/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Connect API
  description: |
    Campus Connect - Social Learning Platform API
    
    Campus Connect is a comprehensive social learning platform for university students featuring:
    - **Study Groups & Squads** - Collaborative learning communities
    - **Learning Sessions** - Scheduled study events and activities
    - **Real-time Chat** - Group messaging and private conversations
    - **Gamification** - Levels, achievements, and leaderboards
    - **Lumina AI** - Intelligent visual learning with animated explanations
    - **Notifications** - Real-time updates for community activity
    
    ## Core Features
    
    ### Social Learning
    - Create and join study groups (squads) with shared interests
    - Organize learning sessions with RSVP tracking
    - Real-time chat rooms for group communication
    - User profiles with topics, interests, and activity tracking
    
    ### Gamification
    - User leveling system (Novice → Learner → Collaborator → Expert → Master)
    - Points earned through engagement (attending events, messaging, creating squads)
    - Achievements and badges
    - Campus-wide leaderboard
    
    ### Lumina AI (Visual Learning)
    Transform natural language queries into educational animations using:
    - **Manim** (3Blue1Brown's animation library)
    - **Gemini 2.5 Flash** (AI script generation)
    - **Daytona.io** (Cloud execution environment)
    - Real-time rendering progress via Server-Sent Events
...
servers:
...
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Create new user account
      description: Register a new user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - username
                - email
                - password
              properties:
                name:
                  type: string
                  example: "John Doe"
                username:
                  type: string
                  example: "johndoe"
                email:
                  type: string
                  format: email
                  example: "john.doe@aun.edu.ng"
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
                studentId:
                  type: string
                  example: "STU12345"
                major:
                  type: string
                  example: "Computer Science"
                year:
                  type: string
                  example: "3"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /groups/{id}:
    get:
      tags:
        - Groups
      summary: Get group details
      description: Retrieve detailed group information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Groups
      summary: Update group
      description: Update group details (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                topics:
                  type: array
                  items:
                    type: string
                location:
                  type: string
                privacy:
                  type: string
                  enum:
                    - public
                    - private
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Groups
      summary: Delete group
      description: Delete group (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Group deleted successfully"
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{id}/join:
    post:
      tags:
        - Groups
      summary: Join group
      description: Join a public or private group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully joined group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Already a member or group is private
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/auth/login:
    post:
      tags:
        - Admin
      summary: Admin login
      description: Authenticate admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: "admin@campusconnect.dev"
                password:
                  type: string
      responses:
        '200':
          description: Admin login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  role:
                    type: string
                    example: "admin"
  /admin/bootstrap:
    post:
      tags:
        - Admin
      summary: Bootstrap admin account
      description: Create initial admin account (first-time setup only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Admin account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /admin/users:
    post:
      tags:
        - Admin
      summary: Create user (admin)
      description: Create a new user account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
                - username
              properties:
                email:
                  type: string
                name:
                  type: string
                username:
                  type: string
                role:
                  type: string
                  enum:
                    - user
                    - admin
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "alice-123"
        name:
          type: string
          example: "Alice Johnson"
        username:
          type: string
          example: "alicejohnson"
        email:
          type: string
          format: email
          example: "alice.johnson@aun.edu.ng"
        role:
          type: string
          enum:
            - user
            - admin
          example: "user"
        avatar:
          type: string
          format: uri
          example: "https://api.dicebear.com/7.x/avataaars/svg?seed=alice"
        bio:
          type: string
          example: "Physics major, love studying quantum mechanics"
        major:
          type: string
          example: "Physics"
        year:
          type: string
          example: "3"
        studentId:
          type: string
          example: "STU12345"
        topics:
          type: array
          items:
            type: string
          example: ["Physics", "Mathematics", "Quantum Computing"]
        createdAt:
          type: string
          format: date-time
    Group:
      type: object
      properties:
        id:
          type: string
          example: "physics-squad-123"
        title:
          type: string
          example: "Physics Study Squad"
        description:
          type: string
          example: "Weekly physics problem-solving sessions"
        ownerId:
          type: string
          example: "alice-123"
        privacy:
          type: string
          enum:
            - public
            - private
          example: "public"
        topics:
          type: array
          items:
            type: string
          example: ["Physics", "Science"]
        location:
          type: string
          example: "Library Room 203"
        memberCount:
          type: integer
          example: 12
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              role:
                type: string
                enum:
                  - owner
                  - moderator
                  - member
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        message:
          type: string
          example: "Missing required field: title"
        code:
          type: string
          example: "VALIDATION_ERROR"
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication using HTTP-only cookies
security:
  - sessionCookie: []